chore: Rename from UserDataSource to InitialDataSource
