+++ data/src/main/java/com/maru/data/datasource/initial/InitialRemoteDataSource.kt
@@ -1,4 +1,4 @@
-package com.maru.data.datasource
+package com.maru.data.datasource.initial
 
 import com.maru.data.model.CoupleInfo
 import com.maru.data.model.User
@@ -7,10 +7,10 @@ import com.maru.data.network.firebase.FirebaseHelper
 import com.maru.data.network.server.RetrofitService
 import javax.inject.Inject
 
-class UserRemoteDataSource @Inject constructor(
+class InitialRemoteDataSource @Inject constructor(
     private val retrofitService: RetrofitService,
     private val firebaseHelper: FirebaseHelper
-) : UserDataSource.Remote {
+) : InitialDataSource.Remote {
 
     override suspend fun registerNewUser(user: RegisterUserRequest): Result<User> = runCatching {
 //        retrofitService.registerNewUser(user)
